+++ codegen/models/ast.py
@@ -213,8 +213,6 @@ class AST:
         return VarScope(self.id, len(self.children))
 
     def _add_stmt(self, stmt: Statement):
-        if isinstance(stmt, NoStatement):
-            return self
         if self._is_frozen:
             raise Exception("The AST is frozen and cannot be modified")
         ast = AST(self.next_child_id(), self.prog, stmt)